chapter_003/src/test/java/ru.job4j/PhoneDictionaryTest.java
package ru.job4j;

import org.junit.Test;
import ru.job4j.search.Person;
import ru.job4j.search.PhoneDictionary;

import java.util.List;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class PhoneDictionaryTest {
    @Test
    public void whenFindByName() {
        PhoneDictionary phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        List<Person> persons = phones.find("Petr");
        assertThat(persons.iterator().next().getSurname(), is("Arsentev"));
    }

    @Test
    public void whenFindByAddress() {
        PhoneDictionary phones = new PhoneDictionary();
        phones.add(
                new Person("Mikhail", "Vyuzhanin", "1234567", "Votkinsk")
        );
        List<Person> persons = phones.find("kinsk");
        assertThat(persons.iterator().next().getAddress(), is("Votkinsk"));
    }
}

